Move viewModels di in one
