--- app/src/main/java/com/lumisdinos/gameoffifteen/di/HomeModule.kt
@@ -1,26 +1,0 @@
-package com.lumisdinos.gameoffifteen.di
-
-import androidx.lifecycle.ViewModel
-import com.lumisdinos.gameoffifteen.di.ViewModelBuilder
-import com.lumisdinos.gameoffifteen.di.ViewModelKey
-import com.lumisdinos.gameoffifteen.presentation.HomeViewModel
-import com.lumisdinos.gameoffifteen.ui.fragment.HomeFragment
-import dagger.Binds
-import dagger.Module
-import dagger.android.ContributesAndroidInjector
-import dagger.multibindings.IntoMap
-
-@Module
-abstract class HomeModule {
-    @ContributesAndroidInjector(
-        modules = [
-            ViewModelBuilder::class
-        ]
-    )
-    internal abstract fun homeFragment(): HomeFragment
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(HomeViewModel::class)
-    abstract fun bindViewModel(viewmodel: HomeViewModel): ViewModel
-}
