--- app/src/main/java/com/lumisdinos/gameoffifteen/di/MainModule.kt
@@ -1,15 +1,0 @@
-package com.lumisdinos.gameoffifteen.di
-
-import com.lumisdinos.gameoffifteen.ui.activity.MainActivity
-import dagger.Module
-import dagger.android.ContributesAndroidInjector
-
-@Module
-abstract class MainModule {
-    @ContributesAndroidInjector(
-        modules = [
-            ViewModelBuilder::class
-        ]
-    )
-    internal abstract fun mainActivity(): MainActivity
-}
